+++ src/model.js/destinations-model.js
@@ -1,0 +1,9 @@
+import { mockDestinations } from '../mocks.js/destination';
+
+export default class PointsModel {
+  destinations = mockDestinations;
+
+  getDestinations() {
+    return this.destinations;
+  }
+}
