+++ src/model.js/offers-model.js
@@ -1,0 +1,10 @@
+import { mockOffers } from '../mocks.js/offers';
+
+
+export default class PointsModel {
+  offers = mockOffers;
+
+  getOffers() {
+    return this.offers;
+  }
+}
