+++ src/view/route-point-view.js
@@ -12,18 +12,18 @@ const createPointTitleTempate = (destination, type) => {
 };
 
 const createPointOffersTemplate = (offersPoint, offers) => {
-  if (offersPoint.offers) {
-    const filteredOffers = offersPoint.offers.filter((offer) => offers.includes(offer.id));
-    return (
-      filteredOffers.map(({title, price}) => (
-        `<li class="event__offer">
-          <span class="event__offer-title">${title}</span>
-          &plus;&euro;&nbsp;
-          <span class="event__offer-price">${price}</span>
-        </li>`)).join('')
-    );
+  if (!offersPoint.offers) {
+    return '';
   }
-  return '';
+  const filteredOffers = offersPoint.offers.filter((offer) => offers.includes(offer.id));
+  return (
+    filteredOffers.map(({title, price}) => (
+      `<li class="event__offer">
+        <span class="event__offer-title">${title}</span>
+        &plus;&euro;&nbsp;
+        <span class="event__offer-price">${price}</span>
+      </li>`)).join('')
+  );
 };
 
 const createPointViewTemplate = (point, offersApp, destinationsApp) => {
